# api/generated/client/typescript-axios/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/games/today": {
      "get": {
        "operationId": "Games_today",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "summary": "List games for today",
        "tags": [
          "games"
        ]
      }
    },
    "/games/id/{id}": {
      "get": {
        "operationId": "Games_findByMyId",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "List games by internal ID",
        "tags": [
          "games"
        ]
      }
    },
    "/games/providerId/{providerGameId}": {
      "get": {
        "operationId": "Games_findByProviderId",
        "parameters": [
          {
            "name": "providerGameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GameViewDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "List games by provider ID",
        "tags": [
          "games"
        ]
      }
    },
    "/games/series/{providerGameId}": {
      "get": {
        "operationId": "Games_getSeries",
        "parameters": [
          {
            "name": "providerGameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeriesDto"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        },
        "summary": "Season series between the two teams in a given game",
        "tags": [
          "games"
        ]
      }
    },
    "/games/season": {
      "get": {
        "operationId": "Games_seasonSchedule",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "season",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SeasonGameDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Full regular-season schedule for a team",
        "tags": [
          "games"
        ]
      }
    },
    "/games/upcoming": {
      "get": {
        "operationId": "Games_upcoming",
        "parameters": [
          {
            "name": "teamId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "count",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameDto"
                  }
                }
              }
            }
          }
        },
        "summary": "Next N scheduled regular-season games for a team (Upcoming tab)",
        "tags": [
          "games"
        ]
      }
    },
    "/games": {
      "get": {
        "operationId": "Games_listByDate",
        "parameters": [
          {
            "name": "date",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/GameViewDto"
                  }
                }
              }
            }
          },
          "500": {
            "description": ""
          }
        },
        "summary": "List games for specific date",
        "tags": [
          "games"
        ]
      }
    },
...
    "/games/{providerGameId}/alerts": {
      "get": {
        "operationId": "Alerts_listAlertsForGame",
        "parameters": [
          {
            "name": "providerGameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Alerts"
        ]
      }
    },
...
    "/games/{gameId}/clips": {
      "get": {
        "operationId": "Clips_forGame",
        "parameters": [
          {
            "name": "gameId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClipDto"
                  }
                }
              }
            }
          }
        },
        "summary": "All of a game's clips, ordered by at-bat index (unmatched ones last).",
        "tags": [
          "clips"
        ]
      }
    },
    "/clips/following": {
      "get": {
        "operationId": "Clips_following",
        "parameters": [
          {
            "name": "players",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "teams",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "summary": "Today's clips for followed players/teams, keyed by entity.",
        "tags": [
          "clips"
        ]
      }
    },
...
  },
  "info": {
    "title": "Baseball Realtime API",
    "description": "API for baseball-realtime",
    "version": "1.0.44",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "LinescoreDto": {
        "type": "object",
        "properties": {
          "away": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/LinescoreTeamDto"
              }
            ]
          },
          "home": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/LinescoreTeamDto"
              }
            ]
          },
          "currentInning": {
            "type": "number",
            "example": 5,
            "nullable": true
          },
          "inningHalf": {
            "type": "string",
            "example": "Top",
            "nullable": true
          },
          "isTopInning": {
            "type": "boolean",
            "example": true,
            "nullable": true
          },
          "outs": {
            "type": "number",
            "example": 2,
            "nullable": true
          }
        }
      },
      "ProbablePitcherDto": {
        "type": "object",
        "properties": {
          "mlbId": {
            "type": "number",
            "nullable": true,
            "example": 663554
          },
          "name": {
            "type": "string",
            "nullable": true,
            "example": "Casey Mize"
          },
          "jerseyNumber": {
            "type": "string",
            "nullable": true,
            "example": "12"
          },
          "pitchHand": {
            "type": "string",
            "nullable": true,
            "enum": [
              "L",
              "R"
            ]
          }
        }
      },
      "StarterStatusDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "confirmed",
              "projected",
              "tbd"
            ]
          },
          "confidence": {
            "type": "string",
            "nullable": true,
            "enum": [
              "High",
              "Medium",
              "Low"
            ]
          },
          "lastStart": {
            "type": "string",
            "nullable": true,
            "example": "Jun 15 vs CHC"
          },
          "basis": {
            "type": "string",
            "nullable": true,
            "example": "On turn behind Verlander, on normal 5 days’ rest."
          }
        },
        "required": [
          "status"
        ]
      },
      "GameDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Internal unique game ID (UUID)",
            "example": "2f1e8d10-4b8a-4e59-a07a-61d9c6c4cf71"
          },
          "providerGameId": {
            "type": "string",
            "description": "External provider-assigned game ID",
            "example": "663937"
          },
          "gameDate": {
            "type": "string",
            "description": "UTC calendar date of the game (YYYY-MM-DD)",
            "example": "2025-10-30"
          },
          "homeAbbr": {
            "type": "string",
            "description": "Home team abbreviation (usually 2–5 chars)",
            "example": "HOU",
            "maxLength": 5
          },
          "awayAbbr": {
            "type": "string",
            "description": "Away team abbreviation (usually 2–5 chars)",
            "example": "DET",
            "maxLength": 5
          },
          "homeName": {
            "type": "string",
            "description": "Home team name",
            "example": "Astros"
          },
          "awayName": {
            "type": "string",
            "description": "Away team name",
            "example": "Tigers"
          },
          "status": {
            "type": "string",
            "description": "Current game status",
            "enum": [
              "scheduled",
              "live",
              "final"
            ],
            "example": "final"
          },
          "startTimeUtc": {
            "type": "string",
            "description": "Start time of the game in UTC (timestamp)",
            "format": "date-time",
            "example": "2025-10-30T19:05:00Z",
            "nullable": true
          },
          "snapshot": {
            "type": "object",
            "description": "Arbitrary JSON snapshot of provider metadata (venue, IDs, etc.)",
            "example": {
              "venue": "Minute Maid Park",
              "providerVenueId": 2143,
              "feedUrl": "https://api.sportsdata.io/v4/mlb/scores/json/Game/663937"
            },
            "nullable": true
          },
          "homeScore": {
            "type": "number",
            "description": "Current home team score",
            "example": 5,
            "nullable": true
          },
          "awayScore": {
            "type": "number",
            "description": "Current away team score",
            "example": 3,
            "nullable": true
          },
          "detailedState": {
            "type": "string",
            "description": "Provider detailed state (e.g., Delayed, Postponed, Suspended). Useful for edge statuses.",
            "example": "Delayed Start",
            "nullable": true
          },
          "inning": {
            "type": "number",
            "description": "Current inning number (live games only)",
            "example": 5,
            "nullable": true
          },
          "half": {
            "type": "string",
            "description": "Current half inning (live games only)",
            "enum": [
              "top",
              "bottom"
            ],
            "example": "top",
            "nullable": true
          },
          "outs": {
            "type": "number",
            "description": "Current outs (live games only)",
            "example": 2,
            "nullable": true
          },
          "linescore": {
            "description": "Full linescore snapshot for the current game state",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/LinescoreDto"
              }
            ]
          },
          "currentInning": {
            "type": "number",
            "description": "Current inning number (alias used by some provider feeds)",
            "example": 5,
            "nullable": true
          },
          "isTopInning": {
            "type": "boolean",
            "description": "Whether it is currently the top half of the inning",
            "example": true,
            "nullable": true
          },
          "halfInning": {
            "type": "string",
            "description": "Half-inning label (e.g. \"Top\", \"Bottom\")",
            "example": "Top",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "description": "Ballpark name",
            "example": "Daikin Park",
            "nullable": true
          },
          "homeTeamId": {
            "type": "number",
            "description": "MLB numeric home team ID",
            "example": 117,
            "nullable": true
          },
          "awayTeamId": {
            "type": "number",
            "description": "MLB numeric away team ID",
            "example": 116,
            "nullable": true
          },
          "homeProbable": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProbablePitcherDto"
              }
            ]
          },
          "awayProbable": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProbablePitcherDto"
              }
            ]
          },
          "homeStarterStatus": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/StarterStatusDto"
              }
            ]
          },
          "awayStarterStatus": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/StarterStatusDto"
              }
            ]
          }
        },
        "required": [
          "gameDate",
          "homeAbbr",
          "awayAbbr",
          "homeName",
          "awayName",
          "status"
        ]
      },
...
      "GameViewDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Internal unique game ID (UUID)",
            "example": "2f1e8d10-4b8a-4e59-a07a-61d9c6c4cf71"
          },
          "providerGameId": {
            "type": "string",
            "description": "External provider-assigned game ID",
            "example": "663937"
          },
          "gameDate": {
            "type": "string",
            "description": "UTC calendar date of the game (YYYY-MM-DD)",
            "example": "2025-10-30"
          },
          "homeAbbr": {
            "type": "string",
            "description": "Home team abbreviation (usually 2–5 chars)",
            "example": "HOU",
            "maxLength": 5
          },
          "awayAbbr": {
            "type": "string",
            "description": "Away team abbreviation (usually 2–5 chars)",
            "example": "DET",
            "maxLength": 5
          },
          "homeName": {
            "type": "string",
            "description": "Home team name",
            "example": "Astros"
          },
          "awayName": {
            "type": "string",
            "description": "Away team name",
            "example": "Tigers"
          },
          "status": {
            "type": "string",
            "description": "Current game status",
            "enum": [
              "scheduled",
              "live",
              "final"
            ],
            "example": "final"
          },
          "startTimeUtc": {
            "type": "string",
            "description": "Start time of the game in UTC (timestamp)",
            "format": "date-time",
            "example": "2025-10-30T19:05:00Z",
            "nullable": true
          },
          "snapshot": {
            "type": "object",
            "description": "Arbitrary JSON snapshot of provider metadata (venue, IDs, etc.)",
            "example": {
              "venue": "Minute Maid Park",
              "providerVenueId": 2143,
              "feedUrl": "https://api.sportsdata.io/v4/mlb/scores/json/Game/663937"
            },
            "nullable": true
          },
          "homeScore": {
            "type": "number",
            "description": "Current home team score",
            "example": 5,
            "nullable": true
          },
          "awayScore": {
            "type": "number",
            "description": "Current away team score",
            "example": 3,
            "nullable": true
          },
          "detailedState": {
            "type": "string",
            "description": "Provider detailed state (e.g., Delayed, Postponed, Suspended). Useful for edge statuses.",
            "example": "Delayed Start",
            "nullable": true
          },
          "inning": {
            "type": "number",
            "description": "Current inning number (live games only)",
            "example": 5,
            "nullable": true
          },
          "half": {
            "type": "string",
            "description": "Current half inning (live games only)",
            "enum": [
              "top",
              "bottom"
            ],
            "example": "top",
            "nullable": true
          },
          "outs": {
            "type": "number",
            "description": "Current outs (live games only)",
            "example": 2,
            "nullable": true
          },
          "linescore": {
            "description": "Full linescore snapshot for the current game state",
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/LinescoreDto"
              }
            ]
          },
          "currentInning": {
            "type": "number",
            "description": "Current inning number (alias used by some provider feeds)",
            "example": 5,
            "nullable": true
          },
          "isTopInning": {
            "type": "boolean",
            "description": "Whether it is currently the top half of the inning",
            "example": true,
            "nullable": true
          },
          "halfInning": {
            "type": "string",
            "description": "Half-inning label (e.g. \"Top\", \"Bottom\")",
            "example": "Top",
            "nullable": true
          },
          "venue": {
            "type": "string",
            "description": "Ballpark name",
            "example": "Daikin Park",
            "nullable": true
          },
          "homeTeamId": {
            "type": "number",
            "description": "MLB numeric home team ID",
            "example": 117,
            "nullable": true
          },
          "awayTeamId": {
            "type": "number",
            "description": "MLB numeric away team ID",
            "example": 116,
            "nullable": true
          },
          "homeProbable": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProbablePitcherDto"
              }
            ]
          },
          "awayProbable": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ProbablePitcherDto"
              }
            ]
          },
          "homeStarterStatus": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/StarterStatusDto"
              }
            ]
          },
          "awayStarterStatus": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/StarterStatusDto"
              }
            ]
          },
          "homeTeamMeta": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TeamMetaDto"
              }
            ]
          },
          "awayTeamMeta": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/TeamMetaDto"
              }
            ]
          }
        },
        "required": [
          "gameDate",
          "homeAbbr",
          "awayAbbr",
          "homeName",
          "awayName",
          "status"
        ]
      },
...
      "SeriesDto": {
        "type": "object",
        "properties": {
          "awayAbbr": {
            "type": "string",
            "example": "HOU"
          },
          "homeAbbr": {
            "type": "string",
            "example": "CHC"
          },
          "awayWins": {
            "type": "number",
            "example": 1
          },
          "homeWins": {
            "type": "number",
            "example": 1
          },
          "games": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SeriesGameDto"
            }
          }
        },
        "required": [
          "awayAbbr",
          "homeAbbr",
          "awayWins",
          "homeWins",
          "games"
        ]
      },
      "SeasonGameDto": {
        "type": "object",
        "properties": {
          "providerGameId": {
            "type": "string",
            "nullable": true
          },
          "gameDate": {
            "type": "string"
          },
          "startTimeUtc": {
            "type": "string",
            "nullable": true
          },
          "isHome": {
            "type": "boolean"
          },
          "oppAbbr": {
            "type": "string"
          },
          "oppName": {
            "type": "string"
          },
          "oppTeamId": {
            "type": "number",
            "nullable": true
          },
          "status": {
            "type": "string",
            "enum": [
              "scheduled",
              "live",
              "final"
            ]
          },
          "detailedState": {
            "type": "string",
            "nullable": true
          },
          "teamScore": {
            "type": "number",
            "nullable": true
          },
          "oppScore": {
            "type": "number",
            "nullable": true
          },
          "winnerName": {
            "type": "string",
            "nullable": true
          },
          "loserName": {
            "type": "string",
            "nullable": true
          },
          "winnerId": {
            "type": "number",
            "nullable": true
          },
          "loserId": {
            "type": "number",
            "nullable": true
          },
          "homeProbableName": {
            "type": "string",
            "nullable": true
          },
          "awayProbableName": {
            "type": "string",
            "nullable": true
          },
          "currentInning": {
            "type": "number",
            "nullable": true
          },
          "halfInning": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "providerGameId",
          "gameDate",
          "isHome",
          "oppAbbr",
          "oppName",
          "status"
        ]
      },
...
      "ClipPlayerTagDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number",
            "example": 657136
          },
          "teamId": {
            "type": "number",
            "example": 111
          },
          "role": {
            "type": "string",
            "example": "batter"
          }
        },
        "required": [
          "id",
          "teamId",
          "role"
        ]
      },
...
      "ClipDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "aaron-civale-in-play-run-s-to-connor-wong"
          },
          "gameId": {
            "type": "string",
            "example": "824703"
          },
          "atBatIndex": {
            "type": "number",
            "nullable": true,
            "example": 57,
            "description": "null when unmatched to a play (§6a)."
          },
          "inning": {
            "type": "number",
            "nullable": true,
            "example": 7
          },
          "half": {
            "type": "string",
            "nullable": true,
            "enum": [
              "top",
              "bottom"
            ]
          },
          "title": {
            "type": "string",
            "example": "Connor Wong's go-ahead double"
          },
          "description": {
            "type": "string",
            "example": "Connor Wong lines a go-ahead double to left field to give the Red Sox a 4-3 lead in the bottom of the 7th inning"
          },
          "durationSec": {
            "type": "number",
            "example": 30
          },
          "mp4Url": {
            "type": "string",
            "example": "https://.../clip_720p.mp4"
          },
          "thumbnailUrl": {
            "type": "string",
            "nullable": true
          },
          "players": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClipPlayerTagDto"
            }
          },
          "scoreAfter": {
            "nullable": true,
            "allOf": [
              {
                "$ref": "#/components/schemas/ClipScoreDto"
              }
            ]
          },
          "publishedAt": {
            "type": "string",
            "example": "2026-09-26T23:41:07Z"
          }
        },
        "required": [
          "id",
          "gameId",
          "title",
          "description",
          "durationSec",
          "mp4Url",
          "players",
          "publishedAt"
        ]
      }
    }
  }
}